clean up main and Country screen
